% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_cluster.R
\name{print.ac_cluster}
\alias{print.ac_cluster}
\title{Imprime resumo de um objeto ac_cluster}
\usage{
\method{print}{ac_cluster}(x, ...)
}
\arguments{
\item{x}{Objeto \code{ac_cluster}.}

\item{...}{Ignorado.}
}
\value{
\code{x} (invisivel).
}
\description{
Imprime resumo de um objeto ac_cluster
}
\keyword{internal}
