% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_qual_codebook.R
\name{ac_qual_load_codebook}
\alias{ac_qual_load_codebook}
\title{Carregar codebook de arquivo YAML}
\usage{
ac_qual_load_codebook(path, ...)
}
\arguments{
\item{path}{Caminho do arquivo \code{.yaml}.}

\item{...}{Ignorado.}
}
\description{
Carregar codebook de arquivo YAML
}
\concept{qualitative}
