% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_qual_codebook.R
\name{ac_qual_save_codebook}
\alias{ac_qual_save_codebook}
\title{Salvar codebook em arquivo YAML}
\usage{
ac_qual_save_codebook(codebook, path = NULL, ...)
}
\arguments{
\item{codebook}{Objeto \code{ac_codebook}.}

\item{path}{Caminho do arquivo \code{.yaml}.}

\item{...}{Ignorado.}
}
\description{
Salvar codebook em arquivo YAML
}
\concept{qualitative}
